+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Conventional Commits.
 - Visual mode with selectable ranges.
 - Kebab-case transformation command for visual selections via `:case kebab`.
 - CamelCase transformation command for visual selections via `:case camel`.
+- snake_case transformation command for visual selections via `:case snake`.
 - PascalCase transformation command for visual selections via `:case pascal`.
 
 ## [0.1.0] - 2025-12-30